♻️ refactor: refactor desktop layout
