--- src/app/chat/(desktop)/features/SessionList.tsx
@@ -1,35 +1,0 @@
-import { DraggablePanelBody } from '@lobehub/ui';
-import { createStyles } from 'antd-style';
-import { memo } from 'react';
-
-import FolderPanel from '@/features/FolderPanel';
-
-import SessionListContent from '../../features/SessionListContent';
-import Header from './SessionHeader';
-
-const useStyles = createStyles(({ stylish, css, cx }) =>
-  cx(
-    stylish.noScrollbar,
-    css`
-      display: flex;
-      flex-direction: column;
-      gap: 2px;
-      padding: 8px 8px 0;
-    `,
-  ),
-);
-
-const Sessions = memo(() => {
-  const { styles } = useStyles();
-
-  return (
-    <FolderPanel>
-      <Header />
-      <DraggablePanelBody className={styles}>
-        <SessionListContent />
-      </DraggablePanelBody>
-    </FolderPanel>
-  );
-});
-
-export default Sessions;
