+++ src/app/chat/(desktop)/index.tsx
@@ -11,7 +11,6 @@ import PageTitle from '../features/PageTitle';
 import ChatHeader from './features/ChatHeader';
 import Conversation from './features/Conversation';
 import SideBar from './features/SideBar';
-import Layout from './layout.desktop';
 
 const Mobile: FC = dynamic(() => import('../(mobile)'), {
   loading: MobileSwitchLoading,
@@ -20,14 +19,14 @@ const Mobile: FC = dynamic(() => import('../(mobile)'), {
 
 const DesktopPage = memo(() => (
   <ResponsiveContainer Mobile={Mobile}>
-    <Layout>
-      <PageTitle />
-      <ChatHeader />
-      <Flexbox flex={1} height={'calc(100% - 64px)'} horizontal>
-        <Conversation />
-        <SideBar />
-      </Flexbox>
-    </Layout>
+    {/*<Layout>*/}
+    <PageTitle />
+    <ChatHeader />
+    <Flexbox flex={1} height={'calc(100% - 64px)'} horizontal>
+      <Conversation />
+      <SideBar />
+    </Flexbox>
+    {/*</Layout>*/}
   </ResponsiveContainer>
 ));
 export default DesktopPage;
