+++ src/app/layout.tsx
@@ -1,18 +1,15 @@
 import { SpeedInsights } from '@vercel/speed-insights/next';
 import { ResolvingViewport } from 'next';
-import { SessionProvider } from 'next-auth/react';
 import { cookies } from 'next/headers';
 import { PropsWithChildren } from 'react';
 import { isRtlLang } from 'rtl-detect';
 
 import Analytics from '@/components/Analytics';
-import { getServerConfig } from '@/config/server';
 import { DEFAULT_LANG, LOBE_LOCALE_COOKIE } from '@/const/locale';
+import AuthProvider from '@/layout/AuthProvider';
+import GlobalLayout from '@/layout/GlobalLayout';
 import GlobalProvider from '@/layout/GlobalProvider';
-import { API_ENDPOINTS } from '@/services/_url';
 import { isMobileDevice } from '@/utils/responsive';
-
-const { ENABLE_OAUTH_SSO = false } = getServerConfig();
 
 const RootLayout = async ({ children }: PropsWithChildren) => {
   const cookieStore = cookies();
@@ -20,16 +17,14 @@ const RootLayout = async ({ children }: PropsWithChildren) => {
   const lang = cookieStore.get(LOBE_LOCALE_COOKIE);
   const direction = isRtlLang(lang?.value || DEFAULT_LANG) ? 'rtl' : 'ltr';
 
-  const content = ENABLE_OAUTH_SSO ? (
-    <SessionProvider basePath={API_ENDPOINTS.oauth}>{children}</SessionProvider>
-  ) : (
-    children
-  );
-
   return (
     <html dir={direction} lang={lang?.value || DEFAULT_LANG} suppressHydrationWarning>
       <body>
-        <GlobalProvider>{content}</GlobalProvider>
+        <GlobalProvider>
+          <AuthProvider>
+            <GlobalLayout>{children}</GlobalLayout>
+          </AuthProvider>
+        </GlobalProvider>
         <Analytics />
         <SpeedInsights />
       </body>
