+++ src/app/market/(desktop)/index.tsx
@@ -9,7 +9,6 @@ import MobileSwitchLoading from '@/features/MobileSwitchLoading';
 
 import AgentCard from '../features/AgentCard';
 import Index from '../index';
-import Layout from './layout.desktop';
 
 const Mobile: FC = dynamic(() => import('../(mobile)'), {
   loading: MobileSwitchLoading,
@@ -18,9 +17,7 @@ const Mobile: FC = dynamic(() => import('../(mobile)'), {
 
 export default memo(() => (
   <ResponsiveContainer Mobile={Mobile}>
-    <Layout>
-      <Index />
-      <AgentCard CardRender={SpotlightCard as FC<SpotlightCardProps>} />
-    </Layout>
+    <Index />
+    <AgentCard CardRender={SpotlightCard as FC<SpotlightCardProps>} />
   </ResponsiveContainer>
 ));
