+++ src/app/market/_layout/Mobile/index.tsx
@@ -1,0 +1,7 @@
+import { PropsWithChildren } from 'react';
+
+const MobileLayout = ({ children }: PropsWithChildren) => {
+  return children;
+};
+
+export default MobileLayout;
