+++ src/app/market/layout.ts
@@ -1,0 +1,8 @@
+import ServerLayout from '@/components/server/ServerLayout';
+
+import Desktop from './_layout/Desktop';
+import Mobile from './_layout/Mobile';
+
+const MarketLayout = ServerLayout({ Desktop, Mobile });
+
+export default MarketLayout;
