+++ src/app/settings/_layout/Desktop/Header.tsx
@@ -1,0 +1,37 @@
+'use client';
+
+import { ChatHeader, ChatHeaderTitle } from '@lobehub/ui';
+import { Tag } from 'antd';
+import { memo } from 'react';
+import { useTranslation } from 'react-i18next';
+import { Flexbox } from 'react-layout-kit';
+
+import { SettingsTabs } from '@/store/global/initialState';
+
+import { useActiveKey } from './useActiveKey';
+
+const Header = memo(() => {
+  const { t } = useTranslation('setting');
+
+  const activeKey = useActiveKey();
+
+  return (
+    <ChatHeader
+      left={
+        <div style={{ paddingLeft: 8 }}>
+          <ChatHeaderTitle
+            title={
+              <Flexbox align={'center'} gap={8} horizontal>
+                {t(`tab.${activeKey}`)}
+
+                {activeKey === SettingsTabs.Sync && <Tag color={'gold'}>{t('tab.experiment')}</Tag>}
+              </Flexbox>
+            }
+          />
+        </div>
+      }
+    />
+  );
+});
+
+export default Header;
