+++ src/app/settings/_layout/Desktop/useActiveKey.ts
@@ -1,0 +1,9 @@
+import { usePathname } from 'next/navigation';
+
+import { SettingsTabs } from '@/store/global/initialState';
+
+export const useActiveKey = () => {
+  const pathname = usePathname();
+
+  return pathname.split('/').at(-1) as SettingsTabs;
+};
