+++ src/app/settings/layout.ts
@@ -1,0 +1,8 @@
+import ServerLayout from '@/components/server/ServerLayout';
+
+import Desktop from './_layout/Desktop';
+import Mobile from './_layout/Mobile';
+
+const SettingsLayout = ServerLayout({ Desktop, Mobile });
+
+export default SettingsLayout;
