--- src/app/settings/sync/layout.tsx
@@ -1,9 +1,0 @@
-import { PropsWithChildren } from 'react';
-
-import { SettingsTabs } from '@/store/global/initialState';
-
-import SettingLayout from '../layout.server';
-
-export default ({ children }: PropsWithChildren) => {
-  return <SettingLayout activeTab={SettingsTabs.Sync}>{children}</SettingLayout>;
-};
