+++ src/app/welcome/_layout/Mobile/Header.tsx
